Changed background of Search to have better Accessibility score
I changed the background of the "Search" button on the Dashboard
to be blue in color. This, combined with the previous change of the
"View details" text color from aquamarine to blue, were made to target
the metric of Accessibility. For the Simulated Mobile measurements by
Lighthouse, the Accessibility score increased from 82 before these changes
to 84 after these changes. For the Desktop measurements by Lighthouse,
the Accessibility score increased from 83 before these changes to
85 after these changes.